Working now, but getting performance errors
