+++ TouchPortalYTMusic/entry.tp
@@ -1,6 +1,6 @@
 {
   "sdk": 6,
-  "version": 231,
+  "version": 232,
   "name": "YouTube Music Desktop plugin",
   "id": "YoutubeMusic",
   "configuration": {
@@ -297,6 +297,14 @@
           "name": "YT Music Playback Shuffle",
           "format": "Shuffle Current playlist",
           "type": "communicate"
+        },
+        {
+          "id": "KillerBOSS.TouchPortal.Plugin.YTMD.Action.CheckConnection",
+          "name": "YT Music Check Connection",
+          "prefix": "plugin",
+          "type": "communicate",
+          "tryInline": true,
+          "format": "Check YTMD connection and refresh state"
         }
       ],
       "events": [
@@ -465,6 +473,12 @@
 		      "type": "text",
 		      "desc": "YT Music Repeat States",
 		      "default": "NONE"
+		    },
+		    {
+		      "id": "KillerBOSS.TouchPortal.Plugin.YTMD.States.ConnectionDebug",
+		      "type": "text",
+		      "desc": "YTMD Connection Debug Info",
+		      "default": "Not checked"
 		    }
       ]
     }
