+++ TouchPortalYTMusic/tp_client.py
@@ -1,0 +1,3 @@
+import TouchPortalAPI
+
+TPClient = TouchPortalAPI.Client("YoutubeMusic")
